Report an error when the firefox executable is not available

+++ start-firefox.sh
@@ -10,6 +10,11 @@ SCRIPTPATH=$(dirname "$(readlink -f $0)")
 # set the prefs location
 PREFS=$SCRIPTPATH/firefox-prefs.js
 
+if [ ! -e $FIREFOX_COMMAND ]; then
+  echo -en "could not find firefox executable: $FIREFOX_COMMAND\n";
+  exit 1;
+fi
+
 if [ -f $PREFS ]; then
   rm -rf $FIREFOX_HOME/*.integrationtest
   $FIREFOX_COMMAND --no-remote -CreateProfile integrationtest
